refactor: sidebar also works in editor mode
